Check for ps
Reverting previous commit as it looks like SailfishOS for some reason doesn't
handle dependencies on binaries and adding check whether ps is installed and if
not, producing error. This is workaround for official Jolla store as I can't
use normal RPM Requires there due to stupid rules in harbour policies.